+++ lib/src/configs/languages/ru.dart
@@ -1,10 +1,5 @@
-import '../../enums/supported_language.dart';
-
-/// Language settings for SupportedLanguage.ru
-class RU {
-  /// The language
-  static const language = SupportedLanguage.ru;
-
+/// Settings for SupportedLanguage.ru
+abstract class RU {
   /// An array of special characters
   static const specialCharacters = [
     'а',