+++ lib/src/enums/unicode_block.dart
@@ -1,0 +1,7 @@
+enum UnicodeBlock {
+  latinSupplement,
+  latinExtendedA,
+  latinExtendedB,
+  latinExtendedC,
+  superscriptsSubscripts,
+}