+++ lib/src/models/package_settings.dart
@@ -1,6 +1,7 @@
 import '../configs/default_settings.dart';
 import '../enums/supported_language.dart';
 import '../enums/text_expansion_format.dart';
+import '../enums/unicode_block.dart';
 import '../utils/reg_exp_utils.dart';
 import '../utils/utils.dart';
 import 'arb_settings.dart';
@@ -13,6 +14,9 @@ class PackageSettings {
 
   /// Whether to replace the base langauge. Defaults to `false`.
   final bool replaceBase;
+
+  /// When [replaceBase] is true, a list of unicode blocks to use. Defaults to all.
+  final List<UnicodeBlock>? unicodeBlocks;
 
   /// A list of languages to generate.
   final List<SupportedLanguage>? languagesToGenerate;
@@ -42,6 +46,7 @@ class PackageSettings {
   PackageSettings({
     required this.inputFilepath,
     required bool? replaceBase,
+    required List<String>? unicodeBlocks,
     required List<String>? languagesToGenerate,
     required bool? useBrackets,
     required String? textExpansionFormat,
@@ -51,12 +56,13 @@ class PackageSettings {
     ARBSettings? arbSettings,
     CSVSettings? csvSettings,
   })  : replaceBase = replaceBase ?? DefaultSettings.replaceBase,
-        languagesToGenerate =
-            Utils.covertSupportedLangugesFromListString(languagesToGenerate),
+        unicodeBlocks = (replaceBase ?? DefaultSettings.replaceBase)
+            ? (Utils.covertUnicodeBlocksFromListString(unicodeBlocks) ?? DefaultSettings.unicodeBlocks)
+            : null,
+        languagesToGenerate = Utils.covertSupportedLangugesFromListString(languagesToGenerate),
         useBrackets = useBrackets ?? DefaultSettings.useBrackets,
         textExpansionFormat =
-            Utils.convertTextExpansionFormatFromString(textExpansionFormat) ??
-                DefaultSettings.textExpansionFormat,
+            Utils.convertTextExpansionFormatFromString(textExpansionFormat) ?? DefaultSettings.textExpansionFormat,
         patternToIgnore = RegExpUtils.combinePatterns(patternsToIgnore),
         arbSettings = arbSettings ?? ARBSettings(),
         csvSettings = csvSettings ?? CSVSettings();
@@ -64,5 +70,5 @@ class PackageSettings {
   /// Returns a String representation of the model.
   @override
   String toString() =>
-      '{inputFilepath: $inputFilepath, replaceBase: $replaceBase, languagesToGenerate: $languagesToGenerate, useBrackets: $useBrackets, textExpansionFormat: $textExpansionFormat, textExpansionRatio: $textExpansionRatio, patternToIgnore: $patternToIgnore, keysToIgnore: $keysToIgnore, arbSettings: $arbSettings, csvSettings: $csvSettings}';
+      '{inputFilepath: $inputFilepath, replaceBase: $replaceBase, unicodeBlocks, $unicodeBlocks, languagesToGenerate: $languagesToGenerate, useBrackets: $useBrackets, textExpansionFormat: $textExpansionFormat, textExpansionRatio: $textExpansionRatio, patternToIgnore: $patternToIgnore, keysToIgnore: $keysToIgnore, arbSettings: $arbSettings, csvSettings: $csvSettings}';
 }