+++ lib/src/services/pseudo_generator.dart
@@ -2,10 +2,13 @@ import 'dart:math';
 
 import 'package:meta/meta.dart';
 
-import '../configs/language_settings.dart';
+import '../configs/default_settings.dart';
 import '../enums/supported_language.dart';
 import '../enums/text_expansion_format.dart';
+import '../enums/unicode_block.dart';
 import '../extensions/string_extensions.dart';
+import '../extensions/supported_language_extensions.dart';
+import 'unicode_fallback.dart';
 
 /// Base pseudolocalization generation logic which can be utilized by file generators
 mixin PseudoGenerator {
@@ -14,12 +17,14 @@ mixin PseudoGenerator {
 
   static String generatePseudoTranslation(
     String baseText, {
+    List<UnicodeBlock>? unicodeBlocks,
     SupportedLanguage? languageToGenerate,
     required bool useBrackets,
     required TextExpansionFormat textExpansionFormat,
     double? textExpansionRate,
     RegExp? patternToIgnore,
   }) {
+    final blocks = unicodeBlocks ?? DefaultSettings.unicodeBlocks;
     final pseudoTextLength = textExpansionRate != null
         ? (baseText.length * textExpansionRate).ceil()
         : pseudotranslationLengthForText(baseText);
@@ -34,6 +39,7 @@ mixin PseudoGenerator {
           textExpansion = generateRandomSpecialCharacters(
             numberOfExpansionCharactersToGenerate,
             language: languageToGenerate,
+            blocks: blocks,
           );
           break;
         case TextExpansionFormat.repeatVowels:
@@ -58,6 +64,7 @@ mixin PseudoGenerator {
             textExpansion = generateRandomSpecialCharacters(
               numberOfExpansionCharactersToGenerate,
               language: languageToGenerate,
+              blocks: blocks,
             );
           }
           break;
@@ -79,9 +86,14 @@ mixin PseudoGenerator {
             onNonMatch: (value) => addSpecialCharactersToText(
               value,
               language: languageToGenerate,
+              blocks: blocks,
             ),
           )
-        : addSpecialCharactersToText(baseText, language: languageToGenerate);
+        : addSpecialCharactersToText(
+            baseText,
+            language: languageToGenerate,
+            blocks: blocks,
+          );
 
     final useExclamationMarks =
         textExpansionFormat == TextExpansionFormat.exclamationMarks;
@@ -137,10 +149,12 @@ mixin PseudoGenerator {
   static String addSpecialCharactersToText(
     String text, {
     required SupportedLanguage? language,
+    required List<UnicodeBlock> blocks,
   }) {
     final sb = StringBuffer();
     final characters = text.split('');
-    final mappingCharacters = mappingCharactersForSupportedLanguage(language);
+    final mappingCharacters =
+        mappingCharactersForSupportedLanguage(language, blocks: blocks);
     final keys = mappingCharacters.keys.toList();
     for (final character in characters) {
       final index = keys.indexOf(character);
@@ -160,6 +174,7 @@ mixin PseudoGenerator {
   static String generateRandomSpecialCharacters(
     int count, {
     required SupportedLanguage? language,
+    required List<UnicodeBlock> blocks,
   }) {
     if (count < 1) {
       return '';
@@ -167,35 +182,46 @@ mixin PseudoGenerator {
 
     final sb = StringBuffer();
     for (var i = 0; i < count; i++) {
-      sb.write(randomSpecialCharacter(language: language));
+      sb.write(randomSpecialCharacter(
+        language: language,
+        blocks: blocks,
+      ));
     }
     return sb.toString();
   }
 
-  /// Returns a random special character for [language].
+  /// Returns a random special character for [language] and [blocks].
   @visibleForTesting
   static String randomSpecialCharacter({
     required SupportedLanguage? language,
+    required List<UnicodeBlock> blocks,
   }) {
-    final specialCharacters = specialCharactersForSupportedLanguage(language);
+    final specialCharacters = specialCharactersForSupportedLanguage(
+      language,
+      blocks: blocks,
+    );
     return specialCharacters[_random.nextInt(specialCharacters.length)];
   }
 
-  /// Returns a list of special characters for [language].
+  /// Returns a list of special characters for [language] and [blocks].
   @visibleForTesting
   static List<String> specialCharactersForSupportedLanguage(
-          SupportedLanguage? language) =>
-      language == null
-          ? LanguageSettings.fallbackSpecialCharacters
-          : LanguageSettings.specialCharacters[language]!;
+    SupportedLanguage? language, {
+    required List<UnicodeBlock> blocks,
+  }) =>
+      language != null
+          ? language.specialCharacters
+          : UnicodeFallback.specialCharacters(blocks: blocks);
 
-  /// Returns mapping characters for [language].
+  /// Returns mapping characters for [language] and [blocks].
   @visibleForTesting
   static Map<String, List<String>> mappingCharactersForSupportedLanguage(
-          SupportedLanguage? language) =>
-      language == null
-          ? LanguageSettings.fallbackMappingCharacters
-          : LanguageSettings.mappingCharacters[language]!;
+    SupportedLanguage? language, {
+    required List<UnicodeBlock> blocks,
+  }) =>
+      language != null
+          ? language.mappingCharacters
+          : UnicodeFallback.mappingCharacters(blocks: blocks);
 
   static const _numberWords = [
     'one',