+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## [0.5.0] - XX/XX/2021
 
+* Added ARB file support.
 * Added repeated vowels, number words and exclamation marks as text expansion formats.
-* Added ability to completely ignore certain keys during text replacement. This is a breaking change as previously line numbers were used.
+* Added ability to completely ignore certain keys during text replacement. This is a minor breaking change as previously line numbers were used.
 * Added ability to use both replace_base and languages_to_generate together.
+* Added Latin Extended B, Latin Extended C and Superscripts and Subscripts.
 * Moved outputFilepath setting to CSVSettings. This is a minor breaking change.
 * Improved code coverage.